requirements.txt
sentence-transformers
faiss-cpu
psycopg2-binary
neo4j
numpy
pytest
lightrag-hku==1.4.8.2  # Pinned for CVE fix (>=1.3.9), stable API